+++ create.ps1
@@ -27,7 +27,8 @@ function Get-AccessToken {
                 }
             }
             Write-Output (Invoke-RestMethod @splatGetTokenParams).access_token
-        } catch {
+        }
+        catch {
             $PSCmdlet.ThrowTerminatingError($_)
         }
     }
@@ -49,7 +50,8 @@ function Resolve-Intus-InplanningError {
         if ($ErrorObject.ErrorDetails) {
             $httpErrorObj.ErrorDetails = $ErrorObject.ErrorDetails
             $httpErrorObj.FriendlyMessage = $ErrorObject.ErrorDetails
-        } elseif ((-not($null -eq $ErrorObject.Exception.Response) -and $ErrorObject.Exception.GetType().FullName -eq 'System.Net.WebException')) {
+        }
+        elseif ((-not($null -eq $ErrorObject.Exception.Response) -and $ErrorObject.Exception.GetType().FullName -eq 'System.Net.WebException')) {
             $streamReaderResponse = [System.IO.StreamReader]::new($ErrorObject.Exception.Response.GetResponseStream()).ReadToEnd()
             if (-not([string]::IsNullOrWhiteSpace($streamReaderResponse))) {
                 $httpErrorObj.ErrorDetails = $streamReaderResponse
@@ -58,7 +60,8 @@ function Resolve-Intus-InplanningError {
         }
         try {
             $httpErrorObj.FriendlyMessage = ($httpErrorObj.FriendlyMessage | ConvertFrom-Json).error_description
-        } catch {
+        }
+        catch {
             #displaying the old message if an error occurs during an API call, as the error is related to the API call and not the conversion process to JSON.
         }
         Write-Output $httpErrorObj
@@ -67,6 +70,13 @@ function Resolve-Intus-InplanningError {
 #endregion
 
 try {
+
+    # Verify if a user must be either [created ] or just [correlated]
+    $accessToken = Get-AccessToken
+    $headers = [System.Collections.Generic.Dictionary[string, string]]::new()
+    $headers.Add('Content-Type', 'application/json')
+    $headers.Add('Authorization', 'Bearer ' + $accessToken)
+
     # Validate correlation configuration
     if ($actionContext.CorrelationConfiguration.Enabled) {
         $correlationField = $actionContext.CorrelationConfiguration.accountField
@@ -79,12 +89,6 @@ try {
             throw 'Correlation is enabled but [accountFieldValue] is empty. Please make sure it is correctly mapped'
         }
 
-        # Verify if a user must be either [created ] or just [correlated]
-        $accessToken = Get-AccessToken
-        $headers = [System.Collections.Generic.Dictionary[string, string]]::new()
-        $headers.Add('Content-Type', 'application/json')
-        $headers.Add('Authorization', 'Bearer ' + $accessToken)
-
         try {
             $splatGetUserParams = @{
                 Uri     = "$($actionContext.Configuration.BaseUrl)/api/users/$correlationValue"
@@ -92,7 +96,8 @@ try {
                 Method  = 'GET'
             }
             $correlatedAccount = Invoke-RestMethod @splatGetUserParams
-        } catch {
+        }
+        catch {
             if (-not($_.ErrorDetails.Message -match '211 - Object does not exist')) {
                 throw "Cannot get user error: [$($_.Exception.Message)]"
             }
@@ -101,7 +106,8 @@ try {
 
     if ($null -ne $correlatedAccount) {
         $action = 'CorrelateAccount'
-    } else {
+    }
+    else {
         $action = 'CreateAccount'
     }
 
@@ -146,7 +152,8 @@ try {
             })
     }
     $outputContext.success = $true
-} catch {
+}
+catch {
     $outputContext.success = $false
     $ex = $PSItem
     if ($($ex.Exception.GetType().FullName -eq 'Microsoft.PowerShell.Commands.HttpResponseException') -or
@@ -154,7 +161,8 @@ try {
         $errorObj = Resolve-Intus-InplanningError -ErrorObject $ex
         $auditMessage = "Could not create or correlate Intus-Inplanning account. Error: $($errorObj.FriendlyMessage)"
         Write-Warning "Error at Line '$($errorObj.ScriptLineNumber)': $($errorObj.Line). Error: $($errorObj.ErrorDetails)"
-    } else {
+    }
+    else {
         $auditMessage = "Could not create or correlate Intus-Inplanning account. Error: $($ex.Exception.Message)"
         Write-Warning "Error at Line '$($ex.InvocationInfo.ScriptLineNumber)': $($ex.InvocationInfo.Line). Error: $($ex.Exception.Message)"
     }